Phase3: dlsym JIT guard, AudioStream m_engine fix + virtual base, SDL excluded from iOS
- HostSys_iOS: Replace forward declaration with dlsym runtime resolution for pthread_jit_write_protect_np
- AudioStream_iOS: Fix _engine -> m_engine variable name mismatch
- AudioStream: Add virtual keyword to GetOutputVolume and SetOutputVolume
- CMakeLists.txt: Exclude SDL-dependent files (usb-pad-sdl-ff, SDLInputSource) from iOS target
- Volume API: Change u32 to float for volume parameters in SPU2 and AudioStream

+++ pcsx2/Host/AudioStream.cpp
@@ -68,7 +68,7 @@ std::unique_ptr<AudioStream> AudioStream::CreateNullStream(u32 sample_rate, u32 
 
 	std::unique_ptr<AudioStream> stream(new AudioStream(sample_rate, params));
 	stream->BaseInitialize(&StereoSampleReaderImpl, false);
-	stream->SetOutputVolume(0);
+	stream->SetOutputVolume(0.0f);
 	return stream;
 }
 
@@ -521,7 +521,7 @@ void AudioStream::SetPaused(bool paused)
 	m_paused = paused;
 }
 
-void AudioStream::SetOutputVolume(u32 volume)
+void AudioStream::SetOutputVolume(float volume)
 {
 	m_volume = volume;
 }
